cloud_functions/update_inegi/requirements.txt
functions-framework==3.*
requests>=2.31.0
pandas>=2.0.0
google-cloud-bigquery>=3.0.0
google-cloud-bigquery-storage>=2.0.0
pyarrow>=12.0.0
db-dtypes>=1.0.0
